Avoid confusing behavior of flatpak info
I have multiple branches of org.gnome.Platform install system-wide,
and non per-user. And flatpak info gives me:

flatpak info org.gnome.Platform
-> not installed
flatpak info --system org.gnome.Platform
-> multiple branches

This confusing behavior comes from the fact that we are querying
multiple locations and are not careful enough to collate the
errors we get properly. This commit changes things so that we
keep querying the next location as long as we get a 'not installed'
error, and we report the first 'multiple branches' error we get.