fdio: Mostly port to new code style
There's one function that did `unlinkat()` in the cleanup section,
not doing that yet.

Note I uncovered a few bugs in a few places where we didn't preserve errno
before doing an `unlinkat()` in error paths in a few cases.

I also tried to prefix a few more error cases with the system call name.