build-update-repo: Remove unwanted deltas before updating summary
The list of deltas is stored in the summary, so we need to delete
them before we generate the summary, not before.

This means there is a short period where people may use the old summary
which references the old, now deleted summaries. However, that is
better than it referencing the deleted deltas forever.