dbus-proxy: Add generated code for the Containers1 interface
This provides identity (although not yet filtering) for
D-Bus connections from a container.

I've put the generated files in common/, because this will likely
need to move into flatpak-run at some point in future: when the
dbus-daemon is able to do message filtering (not just identity)
and servers that persist until signalled by a fd held open by
bwrap, and the portals have been updated to use the new dbus-daemon
APIs, it should be possible to have flatpak-run do the setup.
However, the minimum-viable-product implementation of Containers1 at
<https://bugs.freedesktop.org/show_bug.cgi?id=101354> does not
have these features.

Signed-off-by: Simon McVittie <smcv@collabora.com>