build-update-repo: Don't remove next-to-last deltas
When creating a new commit of an app, we create deltas of it and from
its parent to it, and remove all other deltas. This can be problematic
during the switchover to the new delta, because some active operation
may be using the current delta when its deleted.

We fix this by keeping the parent and grandparent-to-parent deltas
around (but we never *generate* them).