build: Fix issues when symlinks in / point into /var
Set up /var to $appdir/var binding after all the other filesystem
setup has happened. Before we did not do this, so a link from e.g.
/foo to /var/foo would be created in the appdir var rather than the tmpfs
var, which would fail with EEXIST the second time because that directory
is persisted.

In particular, this was problematic on endless os where /var/home is
a symlink.