build: Add --die-with-parent option and drop it by default
We no longer use bwrap --die-with-parent by default, because it has
problems due to the semantics of PR_SET_PDEATHSIG. In particular, it
will kill bwrap as soon as *any* thread in the parent exits, not just
the main one.

This caused weird problems in gnome-builder, as seen in:
 https://bugzilla.gnome.org/show_bug.cgi?id=783950#c14

It should be safe to use this from single-threaded apps though,
so flatpak-builder can still use it. It just has to explicitly
enable it.