Initial work on accessiblity support
This creates a dbus proxy for the a11y bus and sets AT_SPI_BUS_ADDRESS in the
environment to the filtered bus.

The app is only allowed to send a very limited amount of messages to
org.a11y.atspi.Registry on the bus, and nothing else.

This requires a patch to at-spi2-core to read the address from
AT_SPI_BUS_ADDRESS:
     https://github.com/flatpak/freedesktop-sdk-images/blob/1.6/at-spi2-core-address-env-var.patch

Fixes https://github.com/flatpak/flatpak/issues/79